reformat method_interaction for multiway
